refactor: update starter code to match vite react boilerplate
