+++ src/App.jsx
@@ -1,0 +1,11 @@
+import "./App.css";
+
+function App() {
+  return (
+    <div className="App">
+      <h1>LAB | React</h1>
+    </div>
+  );
+}
+
+export default App;
